requirements.txt
langchain
langchain-community
langgraph
langchain-astradb
langchain-groq
dotenv
pillow
transformers
beautifulsoup4